feat: update core models and database context
- Add new DbSets and relationships to Context

- Update DailyProgress model

- Enhance ApplicationUser with additional properties

+++ SoitMed/Models/DailyProgress.cs
@@ -12,7 +12,7 @@ namespace SoitMed.Models
         public int Id { get; set; }
 
         [Required]
-        public int WeeklyPlanId { get; set; }
+        public long WeeklyPlanId { get; set; }
 
         [Required]
         public DateOnly ProgressDate { get; set; }
